[regression] Make SRCDATE/CVSDATE for the fetchers work again. This regressed in r653
PN should always be set for files as it comes from bitbake.conf but as we are close
to a release make the least intrusive change and copy SRCDATE and CVSDATE to the above
line as well. I think the second return is dead code and we should remove it.